-ModelosSuperiores tiene un map en lugar de un vector. Es casi una copia de Modelo1. - Corregidos erroes de clase PPMC. Adaptada a las nuevas estucturas.

+++ trunk/Kaggle/fiuba.datos.ppmc/Modelo0.h
@@ -16,7 +16,7 @@
 
 class Modelo0 {
 public:
-	Modelo0(unsigned primo);
+	Modelo0(unsigned long primo);
 	void agregarPalabra(std::string unNombre);
 	MapaFrecuencia* devolverPalabra(MapaFrecuencia* unaPalabra);
 	virtual ~Modelo0();
